+++ tests/testthat/test-clipboard.R
@@ -1,0 +1,9 @@
+context("clipboard.R")
+
+test_that("clipboard works", {
+  skip_if_not(clipr::clipr_available())
+
+  clipboard_export(antibiotics)
+  expect_identical(antibiotics,
+                   clipboard_import())
+})